refactor: extract shared normalizePath and deduplicate sleep utilities
Extract identical normalizePath() implementations from skills.ts and
prompt-templates.ts into a shared normalize-path.ts utility module.

Replace local sleep/wait helper functions in OAuth modules with inline
`await new Promise(r => setTimeout(r, ms))` to avoid unnecessary code
duplication without creating cross-package dependencies.
