+++ packages/pi-ai/src/utils/oauth/google-antigravity.ts
@@ -115,9 +115,8 @@ async function startCallbackServer(): Promise<CallbackServerInfo> {
 					cancelled = true;
 				},
 				waitForCode: async () => {
-					const sleep = () => new Promise((r) => setTimeout(r, 100));
 					while (!result && !cancelled) {
-						await sleep();
+						await new Promise((r) => setTimeout(r, 100));
 					}
 					return result;
 				},
